Reworked permission management

+++ database/migrations/2023_02_18_202920_add_users_cmseditor.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->smallInteger('cmseditor')->default(0);
+            $table->integer('cmseditor')->default(0);
         });
     }
